Moved KeyedWorkItem and related classes to runners-core
