+++ runners/core-java/src/main/java/org/apache/beam/runners/core/KeyedWorkItem.java
@@ -15,9 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util;
+package org.apache.beam.runners.core;
 
 import org.apache.beam.sdk.util.TimerInternals.TimerData;
+import org.apache.beam.sdk.util.WindowedValue;
 
 /**
  * Interface that contains all the timers and elements associated with a specific work item.
